requirements.txt
ConfigArgParse==1.2
matplotlib~=3.5.2
numpy~=1.23.1
opencv-python~=4.6.0.66
pandas~=1.4.3
Pillow~=10.0.1
scikit-image~=0.19.3
scikit-learn~=1.1.2
scipy~=1.10.0-rc1
torch~=1.13.1
torchvision~=0.13.0
tqdm~=4.64.0
